docs(jwt): added doc block for jwt package
