--- packages/jwt/lib/acl.ts
@@ -1,1 +1,0 @@
-import { JWT } from '.'
