+++ packages/jwt/lib/index.ts
@@ -1,21 +1,5 @@
-import { JWT } from './jwt';
-import { GeneratorOptions } from './common';
-
-export * from './errors/index';
-
-export { JWT, GeneratorOptions };
-export * from './acl';
-
-const instance = new JWT();
-
-export function tokenGenerate(
-  applicationId: string,
-  privateKey: string | Buffer,
-  opts?: GeneratorOptions,
-) {
-  return instance.tokenGenerate(applicationId, privateKey, opts);
-}
-
-export function verifySignature(token: string, privateKey: string | Buffer) {
-  return instance.verifySignature(token, privateKey);
-}
+export * from './errors';
+export * from './tokenGenerate';
+export * from './verifySignature';
+export * from './types';
+export * from './jwt';
