+++ packages/jwt/lib/jwt.ts
@@ -1,93 +1,72 @@
-import { JWTInterface, GeneratorOptions, Claims } from './common';
-import { sign, verify } from 'jsonwebtoken';
-import { v4 as uuidv4 } from 'uuid';
-import {
-  MissingApplicationIdError,
-  MissingPrivateKeyError,
-  InvalidPrivateKeyError,
-  InvalidApplicationIdError,
-} from './errors/index';
-import debug from 'debug';
+import { tokenGenerate, validateOptions } from './tokenGenerate';
+import { verifySignature } from "./verifySignature";
+import { GeneratorOptions, Claims } from './types';
 
-const log = debug('vonage:jwt');
+/**
+ * Interface defining the methods for JWT generation.
+ * @ignore
+ * @deprecated Using the class is no longer needed
+ */
+export interface JWTInterface {
+  /**
+   * Generates a JWT token.
+   *
+   * @param {string} applicationId - The application ID.
+   * @param {string | Buffer} privateKey - The private key for signing the JWT.
+   * @param {GeneratorOptions} [opts] - Optional parameters for token generation.
+   * @returns {void}
+   */
+  tokenGenerate(
+    applicationId: string,
+    privateKey: string | Buffer,
+    opts?: GeneratorOptions
+  ): void;
+}
 
+/**
+ * Class implementing the JWTInterface for JWT operations.
+ * @ignore
+ * @deprecated Using the class is no longer needed
+ */
 export class JWT implements JWTInterface {
+  /**
+   * Generates a JWT token.
+   *
+   * @param {string} applicationId - The application ID.
+   * @param {string | Buffer} privateKey - The private key for signing the JWT.
+   * @param {GeneratorOptions} [opts] - Optional parameters for token generation.
+   * @return {string} - The generated JWT token.
+   */
   tokenGenerate(
     applicationId: string,
     privateKey: string | Buffer,
     opts?: GeneratorOptions,
   ): string {
-    log(`Application id: ${applicationId}`);
-    log(`Private key: ${privateKey}`);
-    if (!applicationId) {
-      throw new MissingApplicationIdError();
-    }
-
-    if (!privateKey) {
-      throw new MissingPrivateKeyError();
-    }
-
-    if (typeof applicationId !== 'string') {
-      throw new InvalidApplicationIdError();
-    }
-
-    if (typeof privateKey !== 'string' && !(privateKey instanceof Buffer)) {
-      throw new InvalidPrivateKeyError();
-    }
-
-    const claims = this.validateOptions(opts);
-    log('Claims', claims);
-    claims.application_id = applicationId;
-
-    return sign(claims, privateKey, {
-      algorithm: 'RS256',
-      header: { typ: 'JWT', alg: 'RS256' },
-    });
+    return tokenGenerate(
+      applicationId,
+      privateKey,
+      opts,
+    )
   }
-
+  /**
+   * Verifies the signature of a JWT token.
+   *
+   * @param {string} jwt - The JWT token to verify.
+   * @param {string | Buffer} privateKey - The private key used for verification.
+   * @return {boolean} - True if the signature is valid, false otherwise.
+   */
   verifySignature(jwt: string, privateKey: string | Buffer): boolean {
-    try {
-      verify(jwt, privateKey, {
-        algorithms: ['RS256'],
-      });
-      return true;
-    } catch (error) {
-      log('Error when verifying token', error);
-    }
-
-    return false;
+    return verifySignature(jwt, privateKey);
   }
 
+
+  /**
+   * Validates the options provided for JWT generation.
+   *
+   * @param {GeneratorOptions} [opts] - The options to validate.
+   * @return {Claims} - The validated claims.
+   */
   private validateOptions(opts?: GeneratorOptions): Claims {
-    const now = parseInt((Date.now() / 1000).toString(), 10);
-
-    const ttl = opts?.ttl || 900;
-    if (opts?.ttl) {
-      delete opts.ttl;
-    }
-
-    const claims: Claims = {
-      ...opts,
-      jti: opts?.jti || uuidv4(),
-      iat: opts?.issued_at || now,
-      exp: now + ttl,
-    };
-
-    if (opts?.subject) {
-      claims.sub = opts.subject;
-    }
-
-    if (opts?.acl) {
-      claims.acl = opts.acl;
-    }
-
-    for (const property in opts) {
-      // eslint-disable-next-line
-      if (opts.hasOwnProperty(property)) {
-        claims[property] = opts[property];
-      }
-    }
-
-    return claims;
+    return validateOptions(opts);
   }
 }
