+++ packages/jwt/lib/types/index.ts
@@ -1,0 +1,3 @@
+export * from './claims';
+export * from './generatorOptions';
+export * from './acl';
