+++ packages/jwt/package.json
@@ -1,6 +1,7 @@
 {
+  "$schema": "https://json.schemastore.org/package.json",
   "name": "@vonage/jwt",
-  "version": "1.6.0",
+  "version": "1.7.0",
   "description": "Vonage JWT package. Creates JWT tokens for Vonage API's",
   "homepage": "https://github.com/vonage/vonage-node-sdk/tree/master/packages/jwt#readme",
   "bugs": {
@@ -11,16 +12,30 @@
     "url": "git+https://github.com/Vonage/vonage-node-sdk.git"
   },
   "license": "Apache-2.0",
-  "author": "Kelly J Andrews <kelly@kellyjandrews.com>",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "contributors": [
+    {
+      "name": "Kelly J Andrews",
+      "email": "kelly@kellyjandrews.com"
+    },
+    {
+      "name": "Chris Tankersley",
+      "email": "chris@ctankersley.com"
+    },
+    {
+      "name": "Chuck \"MANCHUCK\" Reeves",
+      "email": "chuck@manchuck.com"
+    }
+  ],
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "files": [
-    "/dist"
+    "dist/**"
   ],
   "scripts": {
     "build": "npm run clean && npm run compile",
     "clean": "npx shx rm -rf dist tsconfig.tsbuildinfo",
-    "compile": "npx tsc --build --verbose"
+    "compile": "npx tsc --build --verbose",
+    "prepublishOnly": "npm run build"
   },
   "dependencies": {
     "debug": "^4.3.4",
@@ -28,10 +43,10 @@
     "uuid": "^9.0.1"
   },
   "devDependencies": {
+    "@types/jsonwebtoken": "9.0.3",
     "@types/uuid": "^9.0.5"
   },
   "publishConfig": {
     "directory": "dist"
-  },
-  "gitHead": "328f18e5c8a458cb4d06d7955ec2399a6ce6f5d8"
+  }
 }
