+++ packages/jwt/typedoc.json
@@ -1,0 +1,6 @@
+{
+  "$schema": "https://typedoc.org/schema.json",
+  "extends": ["../../typedoc.base.json"],
+  "entryPoints": ["lib/index.ts"],
+  "name": "Vonage JWT"
+}
